snapshots: render as preset-style cards + user rename
Addresses review feedback (style snapshots exactly like presets, reuse logic):

- Extract the preset card into a shared PresetItemDisplay component. Presets and
  snapshots now render through the same card; a subtitle prop switches it to
  snapshot mode (date instead of TOOL/STYLE, no preset-only badges) and a nameSlot
  prop lets the name become an inline input for renaming.
- Snapshots reuse the preset preview pipeline: each is fed in as a preset-shaped
  { id, adjustments: state } item, so its thumbnail is the current image rendered
  with that snapshot's edit state. The preview id embeds createdAt, so overwriting
  a snapshot refreshes its thumbnail and the stale one is revoked for free.
- Presets panel is now a 'Snapshots' section (title + cards) above a 'Presets'
  section (title + cards), matching the requested layout.
- Snapshots are renameable: hover a card for a pencil, or right-click > Rename;
  the name edits inline in place (Enter saves, Esc cancels).
